Finish validation split
