+++ libmamba/include/mamba/validation/update_framework_v0_6.hpp
@@ -13,8 +13,8 @@
 
 #include <nlohmann/json_fwd.hpp>
 
-#include "mamba/core/validate.hpp"
 #include "mamba/validation/keys.hpp"
+#include "mamba/validation/update_framework.hpp"
 
 namespace mamba::validation::v0_6
 {
