+++ libmamba/include/mamba/validation/update_framework_v1.hpp
@@ -12,8 +12,8 @@
 
 #include <nlohmann/json_fwd.hpp>
 
-#include "mamba/core/validate.hpp"
 #include "mamba/validation/keys.hpp"
+#include "mamba/validation/update_framework.hpp"
 
 namespace mamba::fs
 {
