+++ libmamba/src/core/link.cpp
@@ -18,10 +18,10 @@
 #include "mamba/core/menuinst.hpp"
 #include "mamba/core/output.hpp"
 #include "mamba/core/transaction_context.hpp"
-#include "mamba/core/validate.hpp"
 #include "mamba/util/build.hpp"
 #include "mamba/util/environment.hpp"
 #include "mamba/util/string.hpp"
+#include "mamba/validation/tools.hpp"
 
 #ifdef __APPLE__
 #include "mamba/core/util_os.hpp"
