+++ libmambapy/src/libmambapy/bindings/legacy.cpp
@@ -33,10 +33,10 @@
 #include "mamba/core/subdirdata.hpp"
 #include "mamba/core/transaction.hpp"
 #include "mamba/core/util_os.hpp"
-#include "mamba/core/validate.hpp"
 #include "mamba/core/virtual_packages.hpp"
 #include "mamba/specs/version.hpp"
 #include "mamba/util/string.hpp"
+#include "mamba/validation/tools.hpp"
 #include "mamba/validation/update_framework_v0_6.hpp"
 
 #include "bindings.hpp"
